requirements.txt
pandas>=1.3.0
numpy>=1.20.0
ortools>=9.4.1874
matplotlib>=3.5.0
seaborn>=0.11.2
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cors==1.0.1
cryptography==44.0.2
diff-match-patch==20241021
Django==5.2
django-cors-headers==4.7.0
django-import-export==4.3.7
django-unfold==0.57.0
djangorestframework==3.16.0
filelock==3.18.0
future==1.0.0
gevent==25.4.1
greenlet==3.2.1
gunicorn==21.2.0
idna==3.10
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tldextract==5.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
psycopg2-binary>=2.9.9 